+++ custom_components/plugwise/coordinator.py
@@ -180,35 +180,27 @@ class PlugwiseDataUpdateCoordinator(DataUpdateCoordinator[dict[str, GwEntityData
     async def _remove_devices(self, removed_devices: set[str]) -> None:
         """Clean registries when removed devices found."""
         device_reg = dr.async_get(self.hass)
-        device_list = dr.async_entries_for_config_entry(
-            device_reg, self.config_entry.entry_id
-        )
-        # First find the Plugwise via_device
-        gateway_device = device_reg.async_get_device({(DOMAIN, self.api.gateway_id)})
-        if gateway_device is None:
-            return  # pragma: no cover
+        for device_id in removed_devices:
+            device_entry = device_reg.async_get_device({(DOMAIN, device_id)})
+            if device_entry is None:
+                self._firmware_list.pop(device_id, None)
+                LOGGER.warning(
+                    "Failed to remove %s device/zone %s, not present in device_registry",
+                    DOMAIN,
+                    device_id,
+                )
+                continue  # pragma: no cover
 
-        via_device_id = gateway_device.id
-        # Then remove the connected orphaned device(s)
-        for device_entry in device_list:
-            for identifier in device_entry.identifiers:
-                device_id = identifier[1]
-                if (
-                    identifier[0] == DOMAIN
-                    and device_entry.via_device_id == via_device_id
-                    and device_id in removed_devices
-                ):
-                    device_reg.async_update_device(
-                        device_entry.id, remove_config_entry_id=self.config_entry.entry_id
-                    )
-                    self._firmware_list.pop(device_id, None)
-                    LOGGER.debug(
-                        "Removed %s device/zone %s %s from device_registry",
-                        DOMAIN,
-                        device_entry.model,
-                        device_id,
-                    )
-                    break
+            device_reg.async_update_device(
+                device_entry.id, remove_config_entry_id=self.config_entry.entry_id
+            )
+            self._firmware_list.pop(device_id, None)
+            LOGGER.debug(
+                "%s %s %s removed from device_registry",
+                DOMAIN,
+                device_entry.model,
+                device_id,
+            )
 
     async def _update_device_firmware(self, data: dict[str, GwEntityData]) -> None:
         """Detect firmware changes and update the device registry."""
@@ -216,20 +208,27 @@ class PlugwiseDataUpdateCoordinator(DataUpdateCoordinator[dict[str, GwEntityData
             if device_id not in self._firmware_list:
                 continue  # pragma: no cover
             if (new_firmware := device.get(FIRMWARE)) != self._firmware_list[device_id]:
-                await self._update_firmware_in_dr(device_id, new_firmware)
-                self._firmware_list[device_id] = new_firmware
+                updated = await self._update_firmware_in_dr(device_id, new_firmware)
+                if updated:
+                    self._firmware_list[device_id] = new_firmware
 
-    async def _update_firmware_in_dr(self, device_id: str, firmware: str | None) -> None:
+    async def _update_firmware_in_dr(self, device_id: str, firmware: str | None) -> bool:
         """Update device sw_version in device_registry."""
         device_reg = dr.async_get(self.hass)
         device_entry = device_reg.async_get_device({(DOMAIN, device_id)})
         if device_entry is None:
-            return  # pragma: no cover
+            LOGGER.warning(
+                "Failed to update firmware in device_registry, %s device %s not found",
+                DOMAIN,
+                device_id,
+            )
+            return False  # pragma: no cover
 
         device_reg.async_update_device(device_entry.id, sw_version=firmware)
         LOGGER.debug(
-            "Updated device firmware for %s %s %s",
+            "Firmware in device_registry updated for %s %s %s",
             DOMAIN,
             device_entry.model,
             device_id,
         )
+        return True